requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0.0
httpx>=0.24.0
aiosmtplib>=2.0.0
chromadb>=0.4.0
XlsxWriter>=3.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
alembic>=1.10.0 
pydantic-settings>=2.0.0